buildd: take sudo secure_path into account for PATH
When using sudo the PATH gets set to a cofigured secure_path, which is
not quite the same as the default found in /etc/environment. Redefine
default_command_environment() to be the combination of the two with the
secure path taking precendence.

By relying on this behavior, we can set /etc/environment and stop
setting the environment out for each command.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>