buildd: add instance configuration & perform compatibility check
A compatibility tag will allow buildd to evolve in the future, providing
an interface to declare a subsequent update as incompatible with the old
one, forcing the user (or application) to clean the instance before
use.  It is possible that this may never increment if compatibility
is forever maintained, but since it cannot be guaranteed, we define this up
front with an extendable InstanceConfiguration for application-specific
customization.

Using class variables to allow application-specific extensions, read the
instance configuration (if available) from /etc/craft-instance.conf using
InstanceConfiguration.

Set the compatibility tag to "buildd-base-v0" and provide guidance on how
the applications should extend this tag.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>