lxd: add buildd remote support (CRAFT-260)
Add configure_buildd_image_remote() helper to add the buildd
remote and return its name.  Calling applications will call
this as part of their setup code prior to launching
an instance.

Add launcher integration tests for the 16.04 and 18.04 buildd
images.  Xenial is currently unsupported because its
/etc/apt/sources.list does not include the -updates
or -security suites, breaking an assumption that snapd is updated
in the buildd image (it ships 2.0.2).  Further work will be
required to support it properly.

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>