lxd instance: account for command environment (CRAFT-309)
When exec'ing a command through LXD, /etc/environment is not read in
by default as it was when previously using sudo. For now, extend
LXDInstance with the default command environment, which can be
overriden by the execute functions' env= parameter.

- Extend Base definition with 'environment' parameter already present
  in BuilddBase.

- Move _rootify_lxc_command() into LXDInstance and account for the
  default command environment.

- Update LXD launcher to set LXDInstance's environment from the
  existing base configuration.  This way the environment will be
  set for all executed commands.

Future work may change the mechanism used to set the environment
(such as using lxc configuration, or wrapper script).

Signed-off-by: Chris Patterson <chris.patterson@canonical.com>