lxd: pass id for host mount uid mapping
Instead of always using the running user's uid for id mapping, allow
the value to be passed by the application so that the uid of the host
directory owner can be used in situations where the two values are
different. Default to the current user's uid to keep compatibility
with existing behavior.

Signed-off-by: Claudio Matsuoka <claudio.matsuoka@canonical.com>