buildd: Pass on snapd no-CDN configuration if set (#133)
launchpad-buildd sets up some specialized configuration to stop snapd
from using the snap store's CDN when running in Canonical's production
build farm, since there are restrictions in place there on general
internet access but the non-CDN storage endpoint is always allowed.  If
this is in place, then we need to propagate it to containers we create.

I encountered this in an lpcraft build, but I think any of the *craft
tools could in principle hit the same problem when running in Launchpad.
launchpad-buildd normally passes HTTP proxy configuration allowing `snap
install` to work, but that doesn't necessarily apply to downloads
initiated by snapd itself, such as base snap downloads.