bases(buildd): track if instance is properly setup
The instance's setup state is now tracked in the instance config file.
This prevents a scenario where an application is interupted while
setting up the instance. On a subsequent run, craft-providers would
assume the instance was fully setup, which could cause a variety of
errors. Now, if the instance did not fully complete setup and auto-clean
is enabled, the instance will be cleaned and recreated. Otherwise, a
BaseConfigurationError will be raised.

Signed-off-by: Callahan Kovacs <callahan.kovacs@canonical.com>