bases(buildd): track if instance is properly setup (#223)
The instance's setup state is now tracked in the instance config file.
This prevents a scenario where an application is interupted while
setting up the instance. On a subsequent run, craft-providers would
assume the instance was fully setup, which could cause a variety of
errors. Now, if the instance did not fully complete setup and auto-clean
is enabled, the instance will be cleaned and recreated. Otherwise, a
BaseConfigurationError will be raised.

Additional changes:
- bases(instance config): add new field "setup"
- bases(buildd): rename `setup_instance_config()`
- tests: refactor buildd unit tests
- base: update base compatibility tag from `base-v0` to `base-v1`

Signed-off-by: Callahan Kovacs <callahan.kovacs@canonical.com>