feat(bases): make BaseName compatible with craft_platforms
This adds two convenience methods to BaseName to make it compatible
with craft-platforms DistroBase and marks the DistroBase as deprecated
in favour of the craft-platforms DistroBase.

It also updates functions that take a BaseName to take a DistroBase
as well.

This does not set a schedule for when BaseName will be removed.

Signed-off-by: Alex Lowe <alex.lowe@canonical.com>